+++ docker-compose.yaml
@@ -1,12 +1,34 @@
+version: '3.8'
+
 services:
+  web:
+    build: .
+    ports:
+      - "8000:8000"
+    volumes:
+      - ./src:/app/src
+#    environment:
+#      - DATABASE_URL=postgresql://user:password@db:5432/fastapi_db
+    depends_on:
+      db:
+        condition: service_healthy
+    restart: always
+
   db:
-    container_name: doreumung-db
-    image: postgres:15
+    image: postgres:15-alpine
+    volumes:
+      - postgres_data:/var/lib/postgresql/data
     environment:
-      POSTGRES_USER: ${POSTGRES_USER}
-      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD}
-      POSTGRES_DB: ${POSTGRES_DB}
+      - POSTGRES_USER=${POSTGRES_USER}
+      - POSTGRES_PASSWORD=${POSTGRES_PASSWORD}
+      - POSTGRES_DB=${POSTGRES_DB}
     ports:
       - "5432:5432"
-    volumes:
-      - db_data:/var/lib/postgresql/data
+#    healthcheck:
+#      test: ["CMD-SHELL", "pg_isready -U user -d fastapi_db"]
+#      interval: 10s
+#      timeout: 5s
+#      retries: 5
+
+volumes:
+  postgres_data:
