+++ src/travel/repo/place_repo.py
@@ -4,7 +4,7 @@ from fastapi import Depends, HTTPException
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from config.database.connection_async import get_async_session
+from src.config.database.connection_async import get_async_session
 from src.travel.models.place import Place, PlaceUpdate
